Add packet system & URL opening functionality
Added a very basic packet system for reading data from the named pipe.
This seemed sensible as two different messages are now being transmitted
along it and more may come in the future.

It's a bit unclean as every piece of data in a packet is transmitted as
if it were using every allocated byte. This works fine as everything is just strings at the
moment, and it keeps the (de)serialisation very simple.

I did this to support requests to open a URL in the Steam overlay
browser. Now NS2 Assistant can open URLs up to 255 characters long, as
soon as I've brought it up to date to use this new system.