Lazily load steam_api_original.dll
I didn't realise calling LoadLibrary & FreeLibrary within DllMain was a no-no, so now it's lazily loaded whenever it's needed.

If more functions are needed in the future it may be better to switch to linking with a reverse-engineered .lib file. I got this working, but it means more effort if something changes in the original dll. But that's unlikely, and I can probably put most of it in a pre-build command, so we'll see.